backend/requirements.txt
# RedFlag CI — Backend Dependencies (NV)
# Python 3.12+ required

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
mangum==0.18.0

# AWS Services
boto3==1.35.0
botocore==1.35.0

# Security Scanning
trufflehog3==3.0.0
GitPython==3.1.43

# HTTP Client (async registry checks)
httpx==0.27.0

# Data Handling
pydantic==2.9.0
python-dateutil==2.9.0

# YAML for Semgrep rules
PyYAML==6.0.2

# HCL parsing (Terraform)
pyhcl==0.4.5

# Utilities
python-dotenv==1.0.1
